navbar back button and svgs change
