+++ ui/src/scss/_base.scss
@@ -761,6 +761,12 @@ button,
         .disabled {
             text-decoration: none;
         }
+        &.is-no-bg {
+            &:hover,
+            &.is-active {
+                background: transparent !important;
+            }
+        }
     }
     &.button-main /* Deprecated */,
     &.is-main {
