+++ ui/src/scss/_components.scss
@@ -45,6 +45,14 @@
             padding-right: var(--gap-edge);
         }
     }
+    &.is-go-back {
+        .wrap {
+            @include mixins.mobile() {
+                padding-left: 0;
+                padding-right: 0;
+            }
+        }
+    }
     .left,
     .right {
         display: flex;
